stylegan2-ada-pytorch/requirements.txt
torch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
insightface==0.2.1
onnxruntime==1.9.0
pandas==1.2.0
Pillow==8.3.2
ninja==1.10.2.2
os
